==> 4-results-discussion/data-n-analyses/results-analysis/doc/1-community-composition.Rmd <==
---
title: "Biodiversity of macrozoobenthic communities - 1"
date: "2018-02-01"
output:
  html_notebook:
    theme: paper
---

## Community structure & composition
These are the basic analyses of the structure and composition of the zoobenthic communities in Burgas Bay for the thesis. The experimental setup included 3 parts:  
- seagrass/sand preliminary experiment in Sozopol Bay (2012)   
- sandy bottoms in Burgas Bay (2013-2014)  
- seagrass meadows in Burgas Bay (2013-2014)     

These are the 3 datasets on which the same sets of analyses will be run.

This notebook describes the first steps of the analysis, from the import, cleaning & preparation of the data, to the basic descriptive statistics, preliminary analyses and plots.

***

First, set some basic options to control code chunks, evaluation behaviour, etc. (code won't appear in the final report, as it's just setup).
```{r setup, include = FALSE}
library(knitr)

knit_hooks$set(small.mar = function(before, options, envir) {
    if (before) par(mar = c(2, 2, .1, 2))  # smaller margin on top
})

## set the working directory to one directory up (I'm keeping the notebooks in their own subdirectory of the project, doc).
opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())

## set knitr options for knitting code into the report.
opts_chunk$set(cache = TRUE, # save results so that code blocks aren't re-run unless code changes
               autodep = TRUE, # ..or unless a relevant earlier code block changed
               cache.comments = FALSE, # don't re-run if the only thing that changed was the comments
               highlight = TRUE, 
               small.mar = TRUE)

```
  Now, define some subdirectories which will (hopefully) keep the project organized.
```{r workspace_setup}
## print the working directory, just to be on the safe side
paste("You are here: ", getwd())

data.dir <- "data"    # input data files
functions.dir <- "R"  # functions & scripts
save.dir <- "output"  # clean data, output from models & more complex calculations
figures.dir <- "figs" # plots & figures 
```
  Import some necessary libraries.
```{r import_packages, message = FALSE}
library(here)
library(vegan)
library(viridis)
library(tidyverse)
```

  Next, import the custom functions we'll use in the analyses. They're organized in a separate subdirectory, in several files by type of analysis, but we'll import them manually (for now).
```{r import_custom_functions}
source(here(functions.dir, "import_data.R"))
```


  Organize some commonly-used ggplot2 modifications into a more convenient (and less repetitive) format.
```{r custom_ggplot_settings_helpers}
## ggplot settings & things that I keep reusing
theme_set(theme_bw())
# ggplot_theme <- list(
#   theme_bw(),
#   theme(element_text(family = "Times"))
# )

## helper to adjust ggplot text size & avoid repetitions 
text_size <- function(text.x = NULL,
                      text.y = NULL,
                      title.x = NULL,
                      title.y = NULL,
                      legend.text = NULL,
                      legend.title = NULL, 
                      strip.x = NULL, 
                      strip.y = NULL) {
  theme(axis.text.x = element_text(size = text.x),
        axis.text.y = element_text(size = text.y),
        axis.title.x = element_text(size = title.x),
        axis.title.y = element_text(size = title.y),
        legend.text = element_text(size = legend.text), 
        legend.title = element_text(size = legend.title), 
        strip.text.x = element_text(size = strip.x), 
        strip.text.y = element_text(size = strip.y)
        )
  }

# log y/min + 1 transform - useful for species counts/biomass data visualization
log_y_min <- function(y) {
  log(y / min(y[y > 0]) + 1)
}
```
<br>  

#### Macrozoobenthic communities in the sandy sediments of Burgas Bay (2013-2014)
Import the zoobenthic abundance & biomass data. 
The abundances are counts of organisms in each sample; the biomasses - wet weights of organisms in g. NOT expressed as individuals.m^-2^ or g.m^-2^ - it doesn't matter for the calculations; will multiply by the corresponding values whenever needed for more presentable output - tables or plots.
```{r abundance_data_sand_import, message = FALSE}
## import & check the abundance data. Dimensions should be 54 x 205 (5 metadata + 200 species columns); all species columns should be numeric.
(zoo.abnd.sand <- import_zoo_data(data.dir = data.dir, 
                                 zoo.data = "zoo_abnd_sand_orig.csv", 
                                 meta.labels = c("station", "month", "year", "habitat", "replicate")))
```
  
Convert the month and year back to numeric (they remained as character during import); then arrange the data frame - by station (first converting to factor), then year and month. This will be the convention used for all subsequent data frames.
```{r clean_abundance_data_sand}
(zoo.abnd.sand <- zoo.abnd.sand %>% 
  mutate(year = as.numeric(year),
         month = as.numeric(month), 
         station = factor(station, levels = c("Kraimorie", "Chukalya", "Akin", "Sozopol", "Agalina", "Paraskeva"))) %>%
  arrange(station, year, month)
)
```
  
Now on to the biomass! 
```{r biomass_data_sand_import, message = FALSE}
## import & check out the biomass data. 
(zoo.biomass.sand <- import_zoo_data(data.dir = data.dir,
                                    zoo.data = "zoo_biomass_sand_orig.csv",
                                    meta.labels = c("station", "month", "year", "habitat", "replicate")))
# same expected structure as the abundance 
```

```{r clean_biomass_data_sand}
(zoo.biomass.sand <- zoo.biomass.sand %>% 
  
  mutate(year = as.numeric(year), 
         month = as.numeric(month), 
         station = factor(station, levels = c("Kraimorie", "Chukalya", "Akin", "Sozopol", "Agalina", "Paraskeva"))) %>%
  
  arrange(station, year, month))
```
  
Everything looks fine! To be on the safe side, save the cleaned abundance & biomass data frames.
```{r save_clean_data_sand}
write_csv(zoo.abnd.sand, 
          path = here::here(save.dir, "abnd_sand_orig_clean.csv"))
write_csv(zoo.biomass.sand, 
          path = here::here(save.dir, "biomass_sand_orig_clean.csv"))
```

  Now, on with the cleaning.  
  First we'll remove the all-0 columns (= species from the master species list which are not present in the current dataset - don't appear at these particular stations & samplings, etc.). This will simplify later analyses, especially the more complex ones that require a lot of time. 
```{r clean_data_sand_2}
# long data is immensely easier to manipulate within the tidyverse, so we'll do some back-and-forth conversions
(zoo.abnd.sand <- zoo.abnd.sand %>% 
  gather(key = species, value = sp_count, -c(station:replicate)) %>% 
  group_by(species) %>% 
  filter(sp_count > 0) %>%
  spread(species, sp_count, fill = 0) %>%
  arrange(station, year, month))

(zoo.biomass.sand <- zoo.biomass.sand %>% 
  gather(key = species, value = sp_biomass, -c(station:replicate)) %>% 
  group_by(species) %>% 
  filter(sp_biomass > 0) %>%
  spread(species, sp_biomass, fill = 0) %>% 
  arrange(station, year, month))
```
  
Ready to proceed with the analyses themselves! 
  
1. Summary statistics for the community structure & composition
  + the **average abundance** for the whole dataset is `r round(mean(select(zoo.abnd.sand, -c(station:replicate)) %>% rowSums) * 20, 0)` ind.m^-2^, with standard deviation `r round(sd(select(zoo.abnd.sand, -c(station:replicate)) %>% rowSums) * 20, 0)`.
  + the **minimum abundance** is `r min(select(zoo.abnd.sand, -c(station:replicate)) %>% rowSums) * 20` ind.m^-2^, in sample `r paste(unite(zoo.abnd.sand %>% filter(row_number() == which.min(select(zoo.abnd.sand, -c(station:replicate)) %>% rowSums)), "new", c(station, replicate)) %>% select(new))` from `r paste(unite(zoo.abnd.sand %>% filter(row_number() == which.min(select(zoo.abnd.sand, -c(station:replicate)) %>% rowSums)), "new", c(month, year)) %>% select(new))`.
  + the **maximum abundance** is `r format(max(select(zoo.abnd.sand, -c(station:replicate)) %>% rowSums) * 20, scientific = FALSE)` ind.m^-2^, in sample `r paste(unite(zoo.abnd.sand %>% filter(row_number() == which.max(select(zoo.abnd.sand, -c(station:replicate)) %>% rowSums)), "new", c(station, replicate)) %>% select(new))` from `r paste(unite(zoo.abnd.sand %>% filter(row_number() == which.max(select(zoo.abnd.sand, -c(station:replicate)) %>% rowSums)), "new", c(month, year)) %>% select(new))`.
  
  + the **average biomass** for the whole dataset is `r round(mean(select(zoo.biomass.sand, -c(station:replicate)) %>% rowSums) * 20, 2)` g.m^-2^, with standard deviation `r round(sd(select(zoo.biomass.sand, -c(station:replicate)) %>% rowSums) * 20, 2)`.
  + the **minimum biomass** is `r round(min(select(zoo.biomass.sand, -c(station:replicate)) %>% rowSums) * 20, 2)` g.m^-2^, in sample `r paste(unite(zoo.biomass.sand %>% filter(row_number() == which.min(select(zoo.biomass.sand, -c(station:replicate)) %>% rowSums)), "new", c(station, replicate)) %>% select(new))` from `r paste(unite(zoo.biomass.sand %>% filter(row_number() == which.min(select(zoo.biomass.sand, -c(station:replicate)) %>% rowSums)), "new", c(month, year)) %>% select(new))`.
  + the **maximum biomass** is `r round(max(select(zoo.biomass.sand, -c(station:replicate)) %>% rowSums) * 20, 2)` g.m^-2^, in sample `r paste(unite(zoo.biomass.sand %>% filter(row_number() == which.max(select(zoo.biomass.sand, -c(station:replicate)) %>% rowSums)), "new", c(station, replicate)) %>% select(new))` from `r paste(unite(zoo.biomass.sand %>% filter(row_number() == which.max(select(zoo.biomass.sand, -c(station:replicate)) %>% rowSums)), "new", c(month, year)) %>% select(new))`.
  
  + there are `r ncol(select(zoo.abnd.sand, -c(station:replicate)))` taxa overall. Let's check them out to see which are identifed to species level. Unfortunately, this has to be done by counting manually the ones that are not actually species.
```{r count_number_taxa_sand}
names(select(zoo.abnd.sand, -c(station:replicate)))
```
  This list includes some larvae, which are not technically separate taxa or groups (presumably, they belong to one of the other taxa on the list): Decapoda and Polychaeta larvae. There are also Pisces larvae, but since no other fish were found, I think it is better to leave them in, because they represent a new taxonomic group. It probably won't end up influencing any analysis but these summaries here. For the more complicated analyses, such singletons will be filtered out, because they don't bring much information.  
  So, the actual number of taxa (both species and upper-level taxa) is `r ncol(select(zoo.abnd.sand, -c(station:replicate), -starts_with("Decapoda"), -starts_with("Polychaeta")))`.  
  
  + the **average number of taxa** for the dataset is `r round(mean(specnumber(select(zoo.abnd.sand, -c(station:replicate), -starts_with("Decapoda"), -starts_with("Polychaeta")))), 0)`, with standard deviation `r round(sd(specnumber(select(zoo.abnd.sand, -c(station:replicate),  -starts_with("Decapoda"), -starts_with("Polychaeta")))), 0)`.
  + the **minimum number of taxa** is `r min(specnumber(select(zoo.abnd.sand, -c(station:replicate), -starts_with("Decapoda"), -starts_with("Polychaeta"))))`, in sample `r paste(unite(zoo.abnd.sand %>% filter(row_number() == which.min(select(zoo.abnd.sand, -c(station:replicate), -starts_with("Decapoda"), -starts_with("Polychaeta")) %>% specnumber)), "new", c(station, replicate)) %>% select(new))` from `r paste(unite(zoo.abnd.sand %>% filter(row_number() == which.min(select(zoo.abnd.sand, -c(station:replicate), -starts_with("Decapoda"), -starts_with("Polychaeta")) %>% specnumber)), "new", c(month, year)) %>% select(new))`. 
  + the **maximum number of taxa** is `r max(specnumber(select(zoo.abnd.sand, -c(station:replicate), -starts_with("Decapoda"), -starts_with("Polychaeta"))))`, in sample `r paste(unite(zoo.abnd.sand %>% filter(row_number() == which.max(select(zoo.abnd.sand, -c(station:replicate), -starts_with("Decapoda"), -starts_with("Polychaeta")) %>% specnumber)), "new", c(station, replicate)) %>% select(new))` from `r paste(unite(zoo.abnd.sand %>% filter(row_number() == which.max(select(zoo.abnd.sand, -c(station:replicate), -starts_with("Decapoda"), -starts_with("Polychaeta")) %>% specnumber)), "new", c(month, year)) %>% select(new))`.
   
  The **total number of taxa per station** for the whole study period is:
```{r summarize_nb_taxa_sand_by_station}
specnumber(select(zoo.abnd.sand, -c(station:replicate), -starts_with("Decapoda"), -starts_with("Polychaeta")), groups = zoo.abnd.sand$station)
```
    
  We're going to transform the two data frames (abundance and biomass) to long format, because it multiple summaries will be done and I don't want to repeat the code every single time - not to mention that it takes time to compute; it's fine when it's a smallish dataset like mine, but if it were larger, it would cause problems.   
```{r transform_data_frames_sand_long}
zoo.abnd.sand.long <- zoo.abnd.sand %>%
  gather(key = species, value = sp_count, -c(station:replicate))

zoo.biomass.sand.long <- zoo.biomass.sand %>%
  gather(key = species, value = sp_biomass, -c(station:replicate))
```

  The **summarized number of taxa** by station and year is: 
```{r summarize_nb_taxa_sand_by_station_year}
zoo.abnd.sand.long %>%
  # remove taxa that shouldn't really be counted separately towards the total/mean number of taxa 
  filter(!(species == "Decapoda larvae" | species == "Polychaeta larvae")) %>%
  group_by(station, year, month, replicate) %>%
  transmute(total_nb_taxa = specnumber(sp_count)) %>%
  distinct(station, year, total_nb_taxa) %>%
  group_by(station, year) %>% 
  # rounded to the nearest integer, because there is no such thing as half a species 
  summarise(mean_nb_taxa = round(mean(total_nb_taxa)), 
            sd_nb_taxa = round(sd(total_nb_taxa)), 
            nb_replicates = n())
```

  The **summarized abundance** by station and year is: 
```{r summarize_abundance_sand_by_station_year}
## small helper function to summarize abundance/biomass. NB hardcoded grouping variables! (ok, because it's only meant to simplify code a little, not really be portable or reusable anyplace else).

summarize_zoo_data <- function(zoo.data, value.col, per.m2, round.to = 0, ...) {
  # zoo.data - long data frame to be summarized
  # value.col - name of column containing the values we're summarizing
  # per.m2 - multiplication factor, to get abundance/biomass per m2 
  # round - should the output be rounded?  
  # ... - grouping variables (unquoted)
  value.col <- enquo(value.col)
  mean_name <- paste0(quo_name(value.col), "_mean")
  sd_name <- paste0(quo_name(value.col), "_sd")
  group.cols <- enquo(...)
  
  zoo.data %>%
    group_by(!!!group.cols, month, replicate) %>%
    transmute(total := sum(!!value.col) * per.m2) %>%
    distinct(!!!group.cols, total) %>%
    group_by(!!!group.cols) %>%
    summarise(!!mean_name := round(mean(total), round.to), 
              !!sd_name := round(sd(total), round.to), 
              nb_replicates = n())
}

summarize_zoo_data(zoo.abnd.sand.long, value.col = sp_count, per.m2 = 20, round.to = 0)
```
  
  The **summarized biomass** by station and year is: 
```{r summarize_biomass_sand_by_station_year}
summarize_zoo_data(zoo.biomass.sand.long, value.col = sp_biomass, per.m2 = 20, round.to = 2)
```
<br>    

2. Taxonomic composition and structure  
  Import the taxonomic data. This includes all species on my current species list, so it applies to all analyses.
```{r import_taxonomic_data}
# print afterwards, as a way to quickly check the structure and make sure there are no weird import errors. Same goes for every import; don't care about space taken up by printing tables. 
(zoo.taxonomy <- read_csv(here::here(data.dir, "zoo_taxonomy.csv")))
```
  
  Leave only the taxa which were present in the sand samples (abundance > 0 in at least one sample). Here it's done by matching the species names against the column names of the (filtered) abundance dataset, again excluding the polychaete & decapod larvae from the count.
```{r subset_taxonomic_data_sand}
current.taxa.sand <- zoo.taxonomy %>%
  filter(species %in% names(select(zoo.abnd.sand, -starts_with("Decapoda"), -starts_with("Polychaeta"))))
```

  First, let's explore the taxonomic composition of the community: number of taxa per phylum/class, etc. No need to save these plots: they'll remain in the notebook, and only the final versions, to be incorporated in the text, will be exported as files.
```{r explore_taxonomic_structure_sand}
## number of taxa per class
# here, the NAs get sorted to the end (the axis order is reversed, that's why they appear on top). But it's better to have them (barchart from base R, which I used previously, silently drops them) - they are not a mistake and should be present: some taxa have no class at all, others are only identified to an upper taxonomic level.
current.taxa.sand %>% 
  count(class, sort = TRUE) %>%
  ggplot(mapping = aes(x = reorder(class, n), y = n)) + 
    geom_bar(stat = "identity", fill = "skyblue") + 
    coord_flip() + 
    labs(title = "Per class", x = "", y = "Number of taxa") +
    theme_bw()

## number of taxa per phylum
current.taxa.sand %>% 
  count(phylum, sort = TRUE) %>%
  ggplot(mapping = aes(x = reorder(phylum, n), y = n)) + 
    geom_bar(stat = "identity", fill = "skyblue") + 
    coord_flip() + 
    labs(title = "Per phylum", x = "", y = "Number of taxa") +
    theme_bw()

```
  
  The most commonly used larger taxonomic groups in the literature are not class & phylum, but Polychaeta, Mollusca, Crustacea, and Varia (= all the rest). Adding them requires some data transformation.
```{r assign_taxonomic_groups_sand}
# the conditions have to be based on fields that do not contain NAs - or, more precisely, they have to be formulated in a way that DOESN'T return NA as a result of the evaluation. Otherwise the reclassification will result in those groups remaining as NA. This behaviour is caused by the way dplyr and mutate treat NAs; works as I expected in base R, so I didn't notice it before.
current.taxa.sand <- current.taxa.sand %>%
  mutate(group = case_when(
    phylum == "Mollusca" ~ "Mollusca", 
    phylum == "Annelida" & class == "Polychaeta" ~ "Polychaeta", 
    phylum == "Arthropoda" & subclass %in% c("Eumalacostraca", "Thecostraca") ~ "Crustacea", 
    TRUE ~ "Varia"
  ))

current.taxa.sand %>%
  count(group)
```
  Good. Now let's plot the number of taxa per (common) taxonomic group, and save as file.
```{r plot_number_taxa_by_taxonomic_groups_sand}
plot.nb.taxa.by.group <- current.taxa.sand %>% 
  count(group) %>%
  ggplot(mapping = aes(x = reorder(group, n), y = n)) + 
    geom_bar(stat = "identity", fill = "skyblue") + 
    coord_flip() + 
    labs(x = "", y = "Number of taxa") +
    text_size(text.x = rel(1.3), text.y = rel(1.3)) + 
    theme_bw()

ggsave(here::here(figures.dir, "nb_taxa_by_tax_group_sand.png"), 
       plot.nb.taxa.by.group, 
       dpi = 300)
```
  
  **Crustacea** are represented by `r nrow(current.taxa.sand %>% group_by(group) %>% count(family) %>% filter(group == "Crustacea"))` families.  
  **Polychaeta** are represented by `r nrow(current.taxa.sand %>% group_by(group) %>% count(family) %>% filter(group == "Polychaeta"))` families.   
  **Mollusca** are represented by `r nrow(current.taxa.sand %>% group_by(group) %>% count(family) %>% filter(group == "Mollusca"))` families.  
  
  The families with the highest number of species are:
```{r breakdown_nb_taxa_families_sand}
lapply(unique(current.taxa.sand$group), 
       function(x) {
         current.taxa.sand %>%
           group_by(group) %>%
           count(family) %>%
           filter(group == x) %>%
           arrange(desc(n))
       })
```
  NB this is a count of the frequency of occurrence of each family; it would be a real count of the number of species if all taxa were identified to species level. As it is, there are some identified to genus or family level - so this count is only indicative. It really should be checked manually and corrected, if needed.     
  Now let's calculate the contribution of each taxonomic group to the **overall community composition by abundance**.
```{r tax_group_abnd_contribution_overall_sand}
## helper function to summarize taxonomic group contribution to abundance and biomass
summarize_tax_group_contr <- function(zoo.data, tax.data, key.col, value.col, ...) {
  # zoo.data - long data frame of count/biomass data
  # tax.data - data frame (wide) of taxonomic information on the same species
  # key.col - name of column to be used as key for joining the two data frames
  # value.col - name of column in zoo.data containing the values to be summarized
  
  value.col <- enquo(value.col)
  group.vars <- quos(...) # allow for multiple grouping variables 

  zoo.data %>%
    left_join(tax.data, by = key.col) %>%
    group_by(!!!group.vars) %>%
    summarise(total := sum(!!value.col)) %>%
    mutate(proportion = total / sum(total))
}

zoo.abnd.sand.long %>% 
  # exclude these poor larvae again (not that they contribute much, it's < 0.1%, but only to get the same numbers as are already in the text)
  filter(!(species == "Decapoda larvae" | species == "Polychaeta larvae")) %>%
  summarize_tax_group_contr(current.taxa.sand, 
                            key.col = "species", 
                            value.col = sp_count, 
                            group) %>%
  arrange(desc(proportion))
```
  
  Now calculate and plot the **contribution of the taxonomic groups to the abundance by station and year**. 
```{r tax_group_abnd_contrib_st_yr_sand}
tax.props.st.yr.sand <- zoo.abnd.sand.long %>% 
  # exclude the larvae
  filter(!(species == "Decapoda larvae" | species == "Polychaeta larvae")) %>%
  summarize_tax_group_contr(current.taxa.sand, 
                            key.col = "species", 
                            value.col = sp_count, 
                            station, year, group) %>%
  arrange(station, year)

## helper for plotting, because all the repetition is killing me. Don't care if everything within is hardcoded, this is only going to be used here and with this data anyway.
plot_tax_props <- function(tax.props.data, x.labels = NULL, col.palette = "viridis") {
  
  # get/make x-axis labels (stations)  
  if(length(x.labels) > 1) {
    x.labs <- x.labels  
  } else {
    x.labs <- paste0(x.labels, 1:length(unique(tax.props.data$station)))
  }

  ggplot(tax.props.data) + 
    geom_bar(mapping = aes(x = station,
                           y = proportion * 100, 
                           fill = group),
             color = "black", size = 0.15, 
             stat = "identity", position = "stack", inherit.aes = TRUE) +
    scale_fill_viridis(option = col.palette, discrete = TRUE, name = "") +
    scale_x_discrete(labels = x.labs) + 
    labs(x = "Station", y = "%")
}

(plot.tax.props.st.yr.sand <- plot_tax_props(tax.props.st.yr.sand, x.labels = "S",col.palette = "viridis") +
    facet_wrap(~year, nrow = 2) + 
    text_size(text.x = rel(1.3), text.y = rel(1.3), title.x = rel(1.3), legend.text = rel(1.3), strip.x = rel(1.3))
)

## save as file
ggsave(here::here(figures.dir, "tax_gr_props_st_yr_sand.png"), 
       plot.tax.props.st.yr.sand, 
       dpi = 300)
```
    
  Check and plot the species which contribute the most to the abundance at each station.
```{r sp_highest_contrib_abnd_sand}
## helper function to get the top n species with the highest abundance/biomass overall
get_top_sp <- function(zoo.data, nb.sp = 20) {
  # zoo.data - count/biomass data - wide format!
  # nb.sp - number of top-contributing species to select

  zoo.data.long <- zoo.data %>%
    gather(key = species, value = sp_count, -c(station:replicate))
  
  # find the top n species in the dataset
  top.n <- zoo.data.long %>%
    group_by(species) %>%
    summarise(total_count = sum(sp_count)) %>%
    arrange(desc(total_count)) %>%
    top_n(nb.sp)
  
  # subset their actual abundance/biomass at each station from the original (wide) dataset (to keep the order; I'm too lazy to figure out how to best rearrange the long data itself)
  top.n.st <- zoo.data %>%
    select(habitat, station, top.n$species) %>%
    gather(key = species, value = sp_count, -c(station, habitat)) %>%
    mutate(species = factor(species, levels = unique(species))) # convert species to factor to keep this order for plotting

  return(top.n.st)
}

top20.counts.sand.st <- get_top_sp(zoo.abnd.sand, nb.sp = 20)

## another helper for the plot, because I'm sick of writing all these parameters over and over - not that it ended up being that much more concise, but ok, teachable moment

plot_top_n <- function(top.n.sp.data, mapping, labs.legend, lab.y, palette) {
  # top.n.sp.data - data frame (long) of top species' counts/biomasses at the different stations
  # mapping - mappings of the aesthetics
  
  ggplot(top.n.sp.data, mapping) +
    geom_point(size = 2.5, alpha = 0.75) + # make points larger & partially transparent
    scale_color_brewer(palette = palette,  labels = labs.legend) + 
    ylab(lab.y) + 
    coord_flip()
}

plot.most.abnd.sp.sand <- plot_top_n(top20.counts.sand.st, 
                                     aes(x = species, y = log_y_min(sp_count), 
                                         colour = station), 
                                     labs.legend = paste0("S", as.numeric(unique(top20.counts.sand.st$station))), 
                                     lab.y = "Abundance (log(y/min + 1))", 
                                     palette = "Set2"
                                     )

(plot.most.abnd.sp.sand <- plot.most.abnd.sp.sand  + 
    text_size(text.x = rel(1.3), text.y = rel(1.15), legend.text = rel(1.2), legend.title = rel(1.2)) + 
    scale_x_discrete(name = "", limits = rev(levels(top20.counts.sand.st$species))) 
)

## save as file
ggsave(here::here(figures.dir, "most_abnd_sp_st_sand.png"), 
       plot.most.abnd.sp.sand, 
       dpi = 300)
```
    
  Now let's calculate the same way the **overall community composition by biomass**.
```{r tax_group_biomass_contrib_overall_sand}
zoo.biomass.sand.long %>% 
  # exclude the larvae
  filter(!(species == "Decapoda larvae" | species == "Polychaeta larvae")) %>%
  summarize_tax_group_contr(current.taxa.sand, 
                            key.col = "species", 
                            value.col = sp_biomass, 
                            group) %>%
  arrange(desc(proportion))
```
  
  Then calculate and plot the  **contribution of the taxonomic groups to the biomass by station and year**. 
```{r tax_group_biomass_contribution_st_yr_sand}
tax.props.biomass.st.yr.sand <- zoo.biomass.sand.long %>% 
  # exclude the larvae
  filter(!(species == "Decapoda larvae" | species == "Polychaeta larvae")) %>%
  summarize_tax_group_contr(current.taxa.sand, 
                            key.col = "species", 
                            value.col = sp_biomass, 
                            station, year, group) %>%
  arrange(station, year)

# plot this
(plot.tax.props.biomass.st.yr.sand <- plot_tax_props(tax.props.biomass.st.yr.sand, x.labels = "S", col.palette = "viridis") + 
    facet_wrap(~year, nrow = 2) + 
    text_size(text.x = rel(1.3), text.y = rel(1.3), title.x = rel(1.3), legend.text = rel(1.3), strip.x = rel(1.3))
)

## save as file
ggsave(here::here(figures.dir, "tax_gr_props_biomass_st_yr_sand.png"), 
       plot.tax.props.biomass.st.yr.sand, 
       dpi = 300)
```

  Now, for the text itself, combine the proportions of the abundance and the biomass in one single plot. 
```{r composite_plot_tax_gr_props_sand}
## first, combine the proportional abundance and biomass tibbles into a single common (long) one
tax.props.sand.all <- full_join(tax.props.st.yr.sand, 
                                tax.props.biomass.st.yr.sand, 
                                by = c("station", "year", "group")) %>%
  select(-starts_with("total")) %>% 
  gather(key = variable, value = proportion, -station, -year, -group) %>%
  mutate(variable = case_when(variable == "proportion.x" ~ "Abundance", 
                              variable == "proportion.y" ~ "Biomass"))

(plot.composite.tax.gr.props.sand <- plot_tax_props(tax.props.sand.all, x.labels = "S") +
    facet_grid(year ~ variable) +
    theme(legend.position = "top", legend.key.size = unit(0.5, "cm")) +
    text_size(text.x = rel(1.2), text.y = rel(1), title.x = rel(1.2), legend.text = rel(1), strip.x = rel(1.2), strip.y = rel(1.2))
)

ggsave(here::here(figures.dir, "tax_gr_props_sand_composite.png"),
       plot.composite.tax.gr.props.sand,       
       dpi = 300, width = 15, units = "cm")
```
   
  Back to the biomass: check and plot the species which contribute the most to the biomass at each station.
```{r sp_highest_contrib_biomass_sand}
## calculate the total biomass counts by species, then select the 20 with the highest biomass
top20.biomass.sand.st <- get_top_sp(zoo.biomass.sand, nb.sp = 20)

## plot them
plot.highest.biomass.sp.sand <- plot_top_n(top20.biomass.sand.st, 
                                     aes(x = species, y = log_y_min(sp_count), 
                                         colour = station), 
                                     labs.legend = paste0("S", as.numeric(unique(top20.biomass.sand.st$station))), 
                                     lab.y = "Biomass (log(y/min + 1))",
                                     palette = "Set2"
                                     )

(plot.highest.biomass.sp.sand <- plot.highest.biomass.sp.sand  + 
    text_size(text.x = rel(1.3), text.y = rel(1.15), legend.text = rel(1.2), legend.title = rel(1.2)) + 
    scale_x_discrete(name = "", limits = rev(levels(top20.biomass.sand.st$species))) 
)
  
## save as file
ggsave(here::here(figures.dir, "highest_biomass_sp_st_sand.png"), 
       plot.highest.biomass.sp.sand, 
       dpi = 300)
```

***

#### Macrozoobenthic communities in the seagrasses of Burgas Bay (2013-2014)
Now we'll do all of this again, for the seagrass communities.
Import the zoobenthic abundance & biomass data. 
```{r abundance_data_zostera_import, message = FALSE}
## import & check the abundance data. Dimensions should be 32 x 205 (5 metadata + 200 species columns); all species columns should be numeric.
(zoo.abnd.zostera <- import_zoo_data(data.dir = data.dir,
                                     zoo.data = "zoo_abnd_zostera_orig.csv", 
                                     meta.labels = c("station", "month", "year", "habitat", "replicate")))
```
  
Convert the month and year back to numeric (they remained as character during import); then arrange the data frame - by station (first converting to factor), then year and month.
```{r clean_abundance_data_zostera}
(zoo.abnd.zostera <- zoo.abnd.zostera %>% 
   mutate(year = as.numeric(year), 
          month = as.numeric(month), 
          station = factor(station, levels = c("Poda", "Otmanli", "Vromos", "Gradina", "Ropotamo"))) %>%
   arrange(station, year, month))
```
  
Now on to the biomass! 
```{r biomass_data_zostera_import, message = FALSE}
## import & check out the biomass data. 
(zoo.biomass.zostera <- import_zoo_data(data.dir = data.dir,
                                        zoo.data = "zoo_biomass_zostera_orig.csv",
                                        meta.labels = c("station", "month", "year", "habitat", "replicate")))
# same expected structure as the abundance 
```

```{r clean_biomass_data_zostera}
(zoo.biomass.zostera <- zoo.biomass.zostera %>% 
   mutate(year = as.numeric(year),
          month = as.numeric(month), 
          station = factor(station, levels = c("Poda", "Otmanli", "Vromos", "Gradina", "Ropotamo"))) %>%
   arrange(station, year, month))
```
  
Goody! To be on the safe side, save the cleaned abundance & biomass data frames.
```{r save_clean_data_zostera}
write_csv(zoo.abnd.zostera, 
          path = here::here(save.dir, "abnd_zostera_orig_clean.csv"))
write_csv(zoo.biomass.zostera, 
          path = here::here(save.dir, "biomass_zostera_orig_clean.csv"))
```

  On with the cleaning.  
  First we'll remove the species that are absent from the current dataset - don't appear at these particular stations & samplings, etc. This will simplify later analyses, especially the more complex ones that require a lot of time. 
```{r clean_data_zostera_2}
# long data is immensely easier to manipulate within the tidyverse, so we'll do some back-and-forth conversions
(zoo.abnd.zostera <- zoo.abnd.zostera %>% 
  gather(key = species, value = sp_count, -c(station:replicate)) %>% 
  group_by(species) %>% 
  filter(sp_count > 0) %>%
  spread(species, sp_count, fill = 0) %>%
  arrange(station, year, replicate))

(zoo.biomass.zostera <- zoo.biomass.zostera %>% 
  gather(key = species, value = sp_biomass, -c(station:replicate)) %>% 
  group_by(species) %>% 
  filter(sp_biomass > 0) %>%
  spread(species, sp_biomass, fill = 0) %>% 
  arrange(station, year, replicate))
```
  
Ready to proceed with the community structure & composition analyses! 
  
1. Summary statistics for the community structure & composition  
Careful with the multilplication factor when going to ind.m^-2^ & g.m^-2^: seagrass communities were sampled with corers, and the sampling surface is different than in sand communities.  
  + the **average abundance** for the whole dataset is `r format(round(mean(select(zoo.abnd.zostera, -c(station:replicate)) %>% rowSums) * (1/0.0078), 0), scientific = FALSE)` ind.m^-2^, with standard deviation `r format(round(sd(select(zoo.abnd.zostera, -c(station:replicate)) %>% rowSums) * (1/0.0078), 0), scientific = FALSE)`.
  + the **minimum abundance** is `r format(round(min(select(zoo.abnd.zostera, -c(station:replicate)) %>% rowSums) * (1/0.0078), 0), scientific = FALSE)` ind.m^-2^, in sample `r paste(unite(zoo.abnd.zostera %>% filter(row_number() == which.min(select(zoo.abnd.zostera, -c(station:replicate)) %>% rowSums)), "new", c(station, replicate)) %>% select(new))` from `r paste(unite(zoo.abnd.zostera %>% filter(row_number() == which.min(select(zoo.abnd.zostera, -c(station:replicate)) %>% rowSums)), "new", c(month, year)) %>% select(new))`.
  + the **maximum abundance** is `r format(round(max(select(zoo.abnd.zostera, -c(station:replicate)) %>% rowSums) * (1/0.0078), 0), scientific = FALSE)` ind.m^-2^, in sample `r paste(unite(zoo.abnd.zostera %>% filter(row_number() == which.max(select(zoo.abnd.zostera, -c(station:replicate)) %>% rowSums)), "new", c(station, replicate)) %>% select(new))` from `r paste(unite(zoo.abnd.zostera %>% filter(row_number() == which.max(select(zoo.abnd.zostera, -c(station:replicate)) %>% rowSums)), "new", c(month, year)) %>% select(new))`.
  
  + the **average biomass** for the whole dataset is `r round(mean(select(zoo.biomass.zostera, -c(station:replicate)) %>% rowSums) * (1/0.0078), 2)` g.m^-2^, with standard deviation `r round(sd(select(zoo.biomass.zostera, -c(station:replicate)) %>% rowSums) * (1/0.0078), 2)`.
  + the **minimum biomass** is `r round(min(select(zoo.biomass.zostera, -c(station:replicate)) %>% rowSums) * (1/0.0078), 2)` g.m^-2^, in sample `r paste(unite(zoo.biomass.zostera %>% filter(row_number() == which.min(select(zoo.biomass.zostera, -c(station:replicate)) %>% rowSums)), "new", c(station, replicate)) %>% select(new))` from `r paste(unite(zoo.biomass.zostera %>% filter(row_number() == which.min(select(zoo.biomass.zostera, -c(station:replicate)) %>% rowSums)), "new", c(month, year)) %>% select(new))`.
  + the **maximum biomass** is `r round(max(select(zoo.biomass.zostera, -c(station:replicate)) %>% rowSums) * (1/0.0078), 2)` g.m^-2^, in sample `r paste(unite(zoo.biomass.zostera %>% filter(row_number() == which.max(select(zoo.biomass.zostera, -c(station:replicate)) %>% rowSums)), "new", c(station, replicate)) %>% select(new))` from `r paste(unite(zoo.biomass.zostera %>% filter(row_number() == which.max(select(zoo.biomass.zostera, -c(station:replicate)) %>% rowSums)), "new", c(month, year)) %>% select(new))`.
  
  + there are `r ncol(select(zoo.abnd.zostera, -c(station:replicate)))` taxa overall. Let's check them out to see which are identifed to species level. Unfortunately, this has to be done by counting manually the ones that are not actually species.
```{r count_number_taxa_zostera}
names(select(zoo.abnd.zostera, -c(station:replicate)))
```

  We'll exclude the Polychaete larvae from this list (they are not technically separate taxa or groups; presumably they belong to one of the other taxa on the list). Again, it's unlikely this will influence any further analysis.  
  So, the actual number of taxa (both species and upper-level taxa) is `r ncol(select(zoo.abnd.zostera, -c(station:replicate), -starts_with("Polychaeta")))`.  
  
  + the **average number of taxa** for the dataset is `r round(mean(specnumber(select(zoo.abnd.zostera, -c(station:replicate), -starts_with("Polychaeta")))), 0)`, with standard deviation `r round(sd(specnumber(select(zoo.abnd.zostera, -c(station:replicate), -starts_with("Polychaeta")))), 0)`.
  + the **minimum number of taxa** is `r min(specnumber(select(zoo.abnd.zostera, -c(station:replicate), -starts_with("Polychaeta"))))`, in sample `r paste(unite(zoo.abnd.zostera %>% filter(row_number() == which.min(select(zoo.abnd.zostera, -c(station:replicate), -starts_with("Polychaeta")) %>% specnumber)), "new", c(station, replicate)) %>% select(new))` from `r paste(unite(zoo.abnd.zostera %>% filter(row_number() == which.min(select(zoo.abnd.zostera, -c(station:replicate), -starts_with("Polychaeta")) %>% specnumber)), "new", c(month, year)) %>% select(new))`. 
  + the **maximum number of taxa** is `r max(specnumber(select(zoo.abnd.zostera, -c(station:replicate), -starts_with("Polychaeta"))))`, in sample `r paste(unite(zoo.abnd.zostera %>% filter(row_number() == which.max(select(zoo.abnd.zostera, -c(station:replicate), -starts_with("Polychaeta")) %>% specnumber)), "new", c(station, replicate)) %>% select(new))` from `r paste(unite(zoo.abnd.zostera %>% filter(row_number() == which.max(select(zoo.abnd.zostera, -c(station:replicate), -starts_with("Polychaeta")) %>% specnumber)), "new", c(month, year)) %>% select(new))`.
   
  The **total number of taxa per station** for the whole study period is:
```{r summarize_nb_taxa_zostera_by_station}
specnumber(select(zoo.abnd.zostera, -c(station:replicate), -starts_with("Polychaeta")), groups = zoo.abnd.zostera$station)
```
  Again, we're going to transform the two data frames (abundance and biomass) to long format, which is more convenient.   
```{r transform_data_frames_zostera_long}
zoo.abnd.zostera.long <- zoo.abnd.zostera %>%
  gather(key = species, value = sp_count, -c(station:replicate))

zoo.biomass.zostera.long <- zoo.biomass.zostera %>%
  gather(key = species, value = sp_biomass, -c(station:replicate))
```

  The **summarized number of taxa** by station and year is: 
```{r summarize_nb_taxa_zostera_by_station_year}
zoo.abnd.zostera.long %>%
  # remove taxa that shouldn't really be counted separately towards the total/mean number of taxa 
  filter(species != "Polychaeta larvae") %>%
  group_by(station, year, month, replicate) %>%
  transmute(total_nb_taxa = specnumber(sp_count)) %>%
  distinct(station, year, total_nb_taxa) %>%
  group_by(station, year) %>% 
  # rounded to the nearest integer, because there is no such thing as half a species 
  summarise(mean_nb_taxa = round(mean(total_nb_taxa)), 
            sd_nb_taxa = round(sd(total_nb_taxa)), 
            nb_replicates = n())
```

  The **summarized abundance** by station and year is: 
```{r summarize_abundance_zostera_by_station_year}
summarize_zoo_data(zoo.abnd.zostera.long, value.col = sp_count, per.m2 = 1/0.0078, round.to = 0)
```
  
  The **summarized biomass** by station and year is: 
```{r summarize_biomass_zostera_by_station_year}
summarize_zoo_data(zoo.biomass.zostera.long, value.col = sp_biomass, per.m2 = 1/0.0078, round.to = 2)
```
<br>  
  2. Taxonomic composition and structure  
  Leave only the taxa which were present in the zostera samples (abundance > 0 in at least one sample).
```{r subset_taxonomic_data_zostera}
current.taxa.zostera <- zoo.taxonomy %>%
  filter(species %in% names(select(zoo.abnd.zostera, -starts_with("Polychaeta"))))
```

  Explore the taxonomic composition of the community: number of taxa per phylum/class, etc.
```{r explore_taxonomic_structure_zostera}
## number of taxa per class
current.taxa.zostera %>% 
  count(class, sort = TRUE) %>%
  ggplot(mapping = aes(x = reorder(class, n), y = n)) + 
    geom_bar(stat = "identity", fill = "light green") + 
    coord_flip() + 
    labs(title = "Per class", x = "", y = "Number of taxa") +
    theme_bw()

## number of taxa per phylum
current.taxa.zostera %>% 
  count(phylum, sort = TRUE) %>%
  ggplot(mapping = aes(x = reorder(phylum, n), y = n)) + 
    geom_bar(stat = "identity", fill = "light green") + 
    coord_flip() + 
    labs(title = "Per phylum", x = "", y = "Number of taxa") +
    theme_bw()
```
  
  Add in the more commonly used larger taxonomic groups (Polychaeta, Mollusca, Crustacea, and Varia).
```{r assign_taxonomic_groups_zostera}
# the conditions have to be based on fields that do not contain NAs - or, more precisely, they have to be formulated in a way that DOESN'T return NA as a result of the evaluation. Otherwise the reclassification will result in those groups remaining as NA. This behaviour is caused by the way dplyr and mutate treat NAs; works as I expected in base R, so I didn't notice it before.
current.taxa.zostera <- current.taxa.zostera %>%
  mutate(group = case_when(
    phylum == "Mollusca" ~ "Mollusca", 
    phylum == "Annelida" & class == "Polychaeta" ~ "Polychaeta", 
    phylum == "Arthropoda" & subclass %in% c("Eumalacostraca", "Thecostraca") ~ "Crustacea", 
    TRUE ~ "Varia"
  ))

current.taxa.zostera %>%
  count(group)
```
  Now plot the number of taxa per (common) taxonomic group, and save as file.
```{r plot_number_taxa_by_taxonomic_groups_zostera}
plot.nb.taxa.by.group <- current.taxa.zostera %>% 
  count(group) %>%
  ggplot(mapping = aes(x = reorder(group, n), y = n)) + 
    geom_bar(stat = "identity", fill = "light green") + 
    coord_flip() + 
    labs(x = "", y = "Number of taxa") +
    theme(axis.text.x = element_text(size = rel(1.3)), 
          axis.text.y = element_text(size = rel(1.3))) +
    theme_bw()

ggsave(here::here(figures.dir, "nb_taxa_by_tax_group_zostera.png"), 
       plot.nb.taxa.by.group, 
       dpi = 300)
```
  
  **Crustacea** are represented by `r nrow(current.taxa.zostera %>% group_by(group) %>% count(family) %>% filter(group == "Crustacea"))` families.  
  **Polychaeta** are represented by `r nrow(current.taxa.zostera %>% group_by(group) %>% count(family) %>% filter(group == "Polychaeta"))` families.   
  **Mollusca** are represented by `r nrow(current.taxa.zostera %>% group_by(group) %>% count(family) %>% filter(group == "Mollusca"))` families.  
  
  The families with the highest number of species are:
```{r breakdown_nb_taxa_families_zostera}
lapply(unique(current.taxa.zostera$group), 
       function(x) {
         current.taxa.zostera %>%
           group_by(group) %>%
           count(family) %>%
           filter(group == x) %>%
           arrange(desc(n))
       })
```
  NB again, this is a count of the frequency of occurrence of each family; it would be a real count of the number of species if all taxa were identified to species level. As it is, there are some identified to genus or family level - so this count is only indicative. It really should be checked manually and corrected, if needed.  
  Now let's calculate the contribution of each taxonomic group to the **overall community composition by abundance**.
```{r tax_group_abnd_contribution_overall_zostera}
zoo.abnd.zostera.long %>% 
  # exclude these poor larvae again (not that they contribute much, it's < 0.1%, but only to get the same numbers as are already in the text)
  filter(!species == "Polychaeta larvae") %>%
  summarize_tax_group_contr(current.taxa.zostera, 
                            key.col = "species", 
                            value.col = sp_count, 
                            group) %>%
  arrange(desc(proportion))
```
    
  Now calculate and plot the **contribution of the taxonomic groups to the abundance by station and year** in the seagrass communities. 
```{r tax_group_abnd_contrib_st_yr_zostera}
tax.props.st.yr.zostera <- zoo.abnd.zostera.long %>% 
  # exclude the larvae
  filter(!species == "Polychaeta larvae") %>%
  summarize_tax_group_contr(current.taxa.zostera, 
                            key.col = "species", 
                            value.col = sp_count, 
                            station, year, group) %>%
  arrange(station, year)

(plot.tax.props.st.yr.zostera <- plot_tax_props(tax.props.st.yr.zostera, x.labels = "Z") +
    facet_wrap(~year, nrow = 2) +
    text_size(text.x = rel(1.3), text.y = rel(1.3), title.x = rel(1.3), legend.text = rel(1.3), strip.x = rel(1.3))
)

## save as file
ggsave(here::here(figures.dir, "tax_gr_props_st_yr_zostera.png"), 
       plot.tax.props.st.yr.zostera, 
       dpi = 300)
```
    
  Check and plot the species which contribute the most to the abundance at each station.
```{r sp_highest_contrib_abnd_zostera}
top20.counts.zostera.st <- get_top_sp(zoo.abnd.zostera, nb.sp = 20)

plot.most.abnd.sp.zostera <- plot_top_n(top20.counts.zostera.st, 
                                     aes(x = species, y = log_y_min(sp_count), 
                                         colour = station), 
                                     labs.legend = paste0("Z", as.numeric(unique(top20.counts.zostera.st$station))), 
                                     lab.y = "Abundance (log(y/min + 1))", 
                                     palette = "Set2"
                                     )

(plot.most.abnd.sp.zostera <- plot.most.abnd.sp.zostera  + 
    text_size(text.x = rel(1.3), text.y = rel(1.15), legend.text = rel(1.2), legend.title = rel(1.2)) + 
    scale_x_discrete(name = "", limits = rev(levels(top20.counts.zostera.st$species))) 
)

## save as file
ggsave(here::here(figures.dir, "most_abnd_sp_st_zostera.png"), 
       plot.most.abnd.sp.zostera, 
       dpi = 300)
```
  
  Now let's calculate the same way the **overall community composition by biomass**.
```{r tax_group_biomass_contrib_overall_zostera}
zoo.biomass.zostera.long %>% 
  # exclude the larvae
  filter(!species == "Polychaeta larvae") %>%
  summarize_tax_group_contr(current.taxa.zostera, 
                            key.col = "species", 
                            value.col = sp_biomass, 
                            group) %>%
  arrange(desc(proportion))
```
  
  Then calculate and plot the  **contribution of the taxonomic groups to the biomass by station and year**. 
```{r tax_group_biomass_contribution_st_yr_zostera}
tax.props.biomass.st.yr.zostera <- zoo.biomass.zostera.long %>% 
  # exclude the larvae
  filter(!species == "Polychaeta larvae") %>%
  summarize_tax_group_contr(current.taxa.zostera, 
                            key.col = "species", 
                            value.col = sp_biomass, 
                            station, year, group) %>%
  arrange(station, year)

# plot this
(plot.tax.props.biomass.st.yr.zostera <- plot_tax_props(tax.props.biomass.st.yr.zostera, x.labels = "Z") +
    facet_wrap(~year, nrow = 2) + 
    text_size(text.x = rel(1.3), text.y = rel(1.3), title.x = rel(1.3), legend.text = rel(1.3), strip.x = rel(1.3))
)

## save as file
ggsave(here::here(figures.dir, "tax_gr_props_biomass_st_yr_zostera.png"), 
       plot.tax.props.biomass.st.yr.zostera, 
       dpi = 300)
```
  
  Combine the proportions of the abundance and the biomass in one single plot. 
```{r composite_plot_tax_gr_props_zostera}
## combine the proportional abundance and biomass tibbles into a single common (long) one
tax.props.zostera.all <- full_join(tax.props.st.yr.zostera, 
                                tax.props.biomass.st.yr.zostera, 
                                by = c("station", "year", "group")) %>%
  select(-starts_with("total")) %>% 
  gather(key = variable, value = proportion, -station, -year, -group) %>%
  mutate(variable = case_when(variable == "proportion.x" ~ "Abundance", 
                              variable == "proportion.y" ~ "Biomass"))

(plot.composite.tax.gr.props.zostera <- plot_tax_props(tax.props.zostera.all, x.labels = "Z") +
    facet_grid(year ~ variable) +
    theme(legend.position = "top", legend.key.size = unit(0.5, "cm")) +
    text_size(text.x = rel(1.2), text.y = rel(1), title.x = rel(1.2), legend.text = rel(1), strip.x = rel(1.2), strip.y = rel(1.2))
)

ggsave(here::here(figures.dir, "tax_gr_props_zostera_composite.png"),
       plot.composite.tax.gr.props.zostera,       
       dpi = 300, width = 15, units = "cm")
```

  Check and plot the species which contribute the most to the biomass at each station.
```{r sp_highest_contrib_biomass_zostera}
## calculate the total biomass counts by species, then select the 20 with the highest biomass
top20.biomass.zostera.st <- get_top_sp(zoo.biomass.zostera, nb.sp = 20)

## plot them
plot.highest.biomass.sp.zostera <- plot_top_n(top20.biomass.zostera.st, 
                                     aes(x = species, y = log_y_min(sp_count), 
                                         colour = station), 
                                     labs.legend = paste0("Z", as.numeric(unique(top20.biomass.zostera.st$station))), 
                                     lab.y = "Biomass (log(y/min + 1))",
                                     palette = "Set2"
                                     )

(plot.highest.biomass.sp.zostera <- plot.highest.biomass.sp.zostera  + 
    text_size(text.x = rel(1.3), text.y = rel(1.15), legend.text = rel(1.2), legend.title = rel(1.2)) + 
    scale_x_discrete(name = "", limits = rev(levels(top20.biomass.zostera.st$species))) 
)
  
## save as file
ggsave(here::here(figures.dir, "highest_biomass_sp_st_zostera.png"), 
       plot.highest.biomass.sp.zostera, 
       dpi = 300)
```
  
***  
#### Macrozoobenthic communities in the seagrass and sand bottoms of Sozopol Bay (2012)  
These analyses have moved to a new home: "community_composition_SozopolBay_2012.Rmd".
